[refactor] addr -> address 컬럼명 변경으로 인한 도메인 및 객체 변경

+++ src/main/java/com/hanshin/shop/service/CartService.java
@@ -44,7 +44,13 @@ public class CartService {
         return cartMapper.delete(id);
     }
 
-    @Transactional public int deleteAll(Long userId) {
+    @Transactional
+    public int deleteAll(Long userId) {
+        final List<CartVO> all = cartMapper.findAll(userId);
+        log.info("deleteALl {}", all.size());
+        if (all.isEmpty() || all.size() == 0) {
+            throw new IllegalStateException("삭제할 상품이 없습니다.");
+        }
         return cartMapper.deleteAll(userId);
     }
 }
